% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCamera.R
\name{autoCamera}
\alias{autoCamera}
\title{Spectra reconstruction using  CAMERA}
\usage{
autoCamera(x)
}
\arguments{
\item{x}{is a xcms Filled object}
}
\value{
PksAn
}
\description{
uses CAMERA functions to process data. A GUI is used to enter
    CAMERA parameters
}
\details{
Spectra reconstruction is done using the function autoCAMERA()
    that includes the 3 CAMERA functions groupFWHM(), findIsotopes() and
    groupCorr().
    A GUI allows the input of the function parameters.
    Consult CAMERA manual for more information;
    (https://www.bioconductor.org/packages/release/bioc/manuals/CAMERA)
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    data(xSetRef)
    dir.create("./QC")
    RTalign_xset(xSetRef,'loess')
    xsAlign <- group(xsAlign, method= "nearest", mzVsRTbalance= 10,
        mzCheck= mzErrAbs,rtCheck= rtDelta, kNN=10)
    xsFilled <- fillPeaks(xsAlign, method="chrom", expand.mz=0.5 )
    minfrac<-0.3
    ne <- 1
    envir = as.environment(ne)
    assign("minfrac", minfrac, envir)
    autoCamera(xsFilled)
    }
}
\seealso{
\code{\link[CAMERA]{groupCorr}}

\code{\link[CAMERA]{findIsotopes}}

\code{\link[CAMERA]{xsAnnotate}}

\code{\link[CAMERA]{groupFWHM}}
}
