% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasePks_Curated.R
\name{BasePks_Curated}
\alias{BasePks_Curated}
\title{Print BP_EICs chromatograms after curation}
\usage{
BasePks_Curated(x)
}
\arguments{
\item{x}{is the subset data matrix ('BasePks')}
}
\value{
Na
}
\description{
generates a table containing the data of curated peaks.
}
\details{
BasePks_Curated() generates an updated data matrix (BasePksCur)
    with the discarded features removed.
    The final matrix is saved in a tsv file named PksBPs_Curated? which is
    saved within the working directory.
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    data(xSetRef)
    dir.create("./QC")
    RTalign_xset(xSetRef,'loess')
    xsAlign <- group(xsAlign, method= "nearest", mzVsRTbalance= 10,
        mzCheck= mzErrAbs,rtCheck= rtDelta, kNN=10)
    xsFilled <- fillPeaks(xsAlign, method="chrom", expand.mz=0.5 )
    printEICsXset(xSetRef,'corrected')
    minfrac<-0.3
    ne <- 1
    envir = as.environment(ne)
    assign("minfrac", minfrac, envir)
    autoCamera(xsFilled)
    FilterDM(PksAn, xSetRef)
    runtimeStart <- Sys.time()
    CollectBP_EICs(BasePks,'corrected')
    BasePks_Curated(BasePks)
    }
}
