% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasePksCur.R
\docType{data}
\name{BasePksCur}
\alias{BasePksCur}
\title{A data frame showing the output of the workflowafter the curation step}
\format{An object of class data.frame}
\usage{
BasePksCur
}
\description{
A data frame showing the output of the workflowafter the curation step
}
\keyword{datasets}
