% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChromParam.R
\name{ChromParam}
\alias{ChromParam}
\title{Chromatographic Parameters Definition,}
\usage{
ChromParam()
}
\value{
returns a list of input chromatographic parameters
}
\description{
Define and save chromatographic parameters via GUI.
}
\details{
ChromParam() allows the user to input the
    chromatographic parametersrelated to the data set to be
    processed and that will be stored and used in
    the later stages of the framework.
    The values to be entered are the retention time range
    of the data set (rt start and rt end), the maximum
    retention time drift observed and the minimum
    and maximum fullwidth at half maximun (FWHM min and max).
}
\examples{
{ChromParam()
}
}
\seealso{
\code{\link[svDialogs]{dlg_input}}getwd
}
