% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectBP_EICs.R
\name{CollectBP_EICs}
\alias{CollectBP_EICs}
\title{Print BP_EICs chromatograms}
\usage{
CollectBP_EICs(x, y)
}
\arguments{
\item{x}{is the subset data matrix ('BasePks') created via the FilterDM
function}

\item{y}{is a character vector ( 'corrected' for branch b and
'filled' for branch a)}
}
\value{
runtime
}
\description{
prints Base Peaks chromatograms (.png)  of the filtered
    features and stores them in a folder
}
\details{
CollectBP_EICs() prints the extracted ion chromatograms related to
    the base peak matrix in a .png format.The pngs files are duplicated into
    2 directories,named EICs_BasePeaks and EICs_BasePeaks_Curated.
    The former is used as data back-up, while the latter allows the user
    to perform a final curation on the EICs,
    by removing those that do not contain useful information
    (e.g. background noise)
}
\examples{
{  ChromParam()
    MassSpecParam()
    PeakPickingParam()
    data(xSetRef)
    dir.create("./QC")
    RTalign_xset(xSetRef,'loess')
    xsAlign <- group(xsAlign, method= "nearest", mzVsRTbalance= 10,
        mzCheck= mzErrAbs,rtCheck= rtDelta, kNN=10)
    xsFilled <- fillPeaks(xsAlign, method="chrom", expand.mz=0.5 )
    printEICsXset(xSetRef,'corrected')
    minfrac<-0.3
    ne <- 1
    envir = as.environment(ne)
    assign("minfrac", minfrac, envir)
    autoCamera(xsFilled)
    FilterDM(PksAn, xSetRef)
    runtimeStart <- Sys.time()
    CollectBP_EICs(BasePks,'corrected')
    }
}
