% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDM.R
\name{CreateDM}
\alias{CreateDM}
\title{Creates a Summary Datamatrix (XCMSnEXP objects)}
\usage{
CreateDM(x, y)
}
\arguments{
\item{x}{is a fillChromPeaks object}

\item{y}{indicates the type of response to be stored ('into', 'maxo')}
}
\value{
Na
}
\description{
creata a summary data matrix containing feature id, mz ,rt and
    intensity value detected for each feature.
}
\details{
peculiar of 'branch a', the function collates features information
    generated by the xcms featureDefinition() function with
    their corresponding response
    (either the integrated peak area or maximum intensity).
    The arguments of the function are the filled XCMSnEXP object and the type
    of instrumental response desired (either into or maxo).
    The data matrix is written into a.cdv file in the QC folder.
}
\examples{
{data(xdata)
    dir.create("./QC")
    xdata <- groupChromPeaks(xdata,
        param = PeakDensityParam(sampleGroups = xdata$sample_group,
        minFraction = 0.3, bw = 20))
    xfilled <- fillChromPeaks(xdata, param =(FillChromPeaksParam(ppm = 50,
        expandMz = 0.5)))
    CreateDM(xfilled, 'maxo')
    }
}
\seealso{
\code{\link[xcms]{XCMSnExp-class}}
}
