% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineClassAttributes.R
\name{DefineClassAttributes}
\alias{DefineClassAttributes}
\title{Define Graphical Attributes}
\usage{
DefineClassAttributes(x)
}
\arguments{
\item{x}{is a character vector containing the sample class names,
as per the ClassType object.}
}
\value{
returns a list of class identifiers (sample groups, color codes and
    numbers corresponding to pch symbols)
}
\description{
Define and stores  colors, pch and samplegroup identifiers
}
\details{
The argument of the function is the object ClassType.
    The function defines and returns specific symbols and colours for each
    sample class, which are then used to produce graphical outputs.
    Sample classesare organised according to the sub-directries present in
    the MSfiles folder.
}
\examples{
{ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    }
}
