% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expRefData.R
\name{expRefData}
\alias{expRefData}
\title{Peak Detection Evaluation on the Package Dataset}
\usage{
expRefData(x)
}
\arguments{
\item{x}{is an OnDiskMSnExp object created using the}
}
\value{
Na
}
\description{
Define parameters and performs peak detection.
    on the reference file dataset (rawDataSet).
}
\details{
First, a GUI allows to define the settings of xcms functions
    findChromPeaks and groupChromPeaks which perform peak picking
    Secondly, data files are read using the
    xcms function readMSData.
    Lastly, peak picking and grouping is carried out.
    The output of the function is the XCMSnEXP object x_refs.
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    SampleGroup<-c("Mix", "Mix")
    symbol<-c(1,1)
    ClassCol<-c("#FF0000FF", "#FF0000FF")
    path<-system.file("extdata",package = "MStractor")
    files <- dir(path, pattern = ".mzXML", full.names = TRUE)
    test<-files[1:2]
    pd <- data.frame(sample_name = sub(basename(test), pattern = filetype,
        replacement ="", fixed = TRUE),sample_group = SampleGroup,
        stringsAsFactors = FALSE)
    ref_data <- readMSData(test, pdata = new("NAnnotatedDataFrame",
        pd), mode = "onDisk")
    expRefData(ref_data)
}
}
\seealso{
\code{\link[xcms]{findChromPeaks-centWave}}
}
