% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterDM.R
\name{FilterDM}
\alias{FilterDM}
\title{Create Summary Data Matrix}
\usage{
FilterDM(x, y)
}
\arguments{
\item{x}{is a peak list form a xsAnnotate object}

\item{y}{a xset object}
}
\value{
a tab separated file is stored within the working directory
}
\description{
filter the feature data matrix by retaining only the most
    intense feature for each group.
}
\examples{
{data(xSetRef)
    data(PksAnDF)
    FilterDM(PksAnDF, xSetRef)
}
}
\seealso{
\code{\link[xcms]{XCMSnExp-class}}
}
