% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTIC.R
\name{getTIC}
\alias{getTIC}
\title{get TIC of a single analysis}
\usage{
getTIC(file, rtcor = NULL)
}
\arguments{
\item{file}{is a file of the analytical set, rtcor defines whether the
rt correction was performed.}

\item{rtcor}{defines whether the rt correction was performed}
}
\value{
Na
}
\description{
function to get a singleTIc
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    data(testX)
    expRefData(testDataSet)}
}
