% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTICs.R
\name{getTICs}
\alias{getTICs}
\title{Prints overlaid TICs for xcmsSet Objects}
\usage{
getTICs(
  xcmsSet = NULL,
  files = NULL,
  pngName = "TICs.png",
  rt = c("raw", "corrected")
)
}
\arguments{
\item{xcmsSet}{is a xcmsSet object (either aligned or not )}

\item{files}{a list of files to print the TICs for}

\item{pngName}{defines the data file name}

\item{rt}{is a vector defining the type of TICs ('raw' or ' corrected')}
}
\value{
Na
}
\description{
function to print a .png file with the overlaid TIcs
}
\details{
the graphical output is stored as .png in theQC folder.
}
\examples{
{ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    data(xdata)
    xsetConvert(xdata)
    sampnames(xset)<-spn
    dir.create('./QC')
    getTICs(xcmsSet= xset, pngName= "./QC/TICs_raw.png", rt= "raw")
}
}
