% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Load Data Set}
\usage{
LoadData()
}
\value{
returns the the list of loaded raw data files.
}
\description{
Loads raw data files form the MsFiles directory and defines
    sample classes.
}
\details{
The function does not  require arguments and returns a complete
    list of the files loaded. It also starts recording the time necessary for
    data processing and creates the object ClassType which associates every
    datafile with the class of belonging.
}
