% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedianNormalize.R
\name{MedianNormalize}
\alias{MedianNormalize}
\title{Matrix Median Normalisation}
\usage{
MedianNormalize(x, y)
}
\arguments{
\item{x}{is the BasePksCur matrix}

\item{y}{is the xcmsSet}
}
\value{
A normalised datamatrix (NormalizedMatrix) as .csv.
}
\description{
Generate a table containing the bormalised data.
}
\examples{
{data(xSetRef)
    data(BasePksCur)
    MedianNormalize(BasePksCur,xSetRef)
}
}
