% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverlaidTICs.R
\name{OverlaidTICs}
\alias{OverlaidTICs}
\title{Print Overlaid TICs for XCMSnEXP objects.}
\usage{
OverlaidTICs(x, y)
}
\arguments{
\item{x}{is a xdata object}

\item{y}{is a character vector containing the type of TICs to be printed
(possible values are 'raw' and 'corrected')}
}
\value{
Na
}
\description{
prints overlaid TICs of the dataset.
}
\details{
the graphical output is stored as .png in theQC folder.
}
\examples{
{ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    dir.create('./QC')
    data(xdata)
    OverlaidTICs(xdata, 'raw')
    }
}
\seealso{
\code{\link[xcms]{chromatogram-method}}
}
