% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakPickGroup.R
\name{peakPickGroup}
\alias{peakPickGroup}
\title{Dataset Peak Detection and Grouping}
\usage{
peakPickGroup()
}
\value{
Na
}
\description{
Define peak picking parameters via GUI and performs peak
    detection and grouping on the entire dataset.
}
\details{
Automatically creates a XCMSnEXP dataset containing all the
    datafiles.The function performs the same steps described for
    exploreRefs(),with the only difference that the processing is applied to
    the whole dataset.
    It is important that the input parameters used in this
    step match the onesdefined for the reference files.
    After performing peak picking on each datafile,
    peaks are matched across all samples and grouped using the xcms
    functions findChromPeaks and groupChromPeaks.
    Results are stored in the XCMSnEXP object xdata.
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    path<-system.file("extdata",package = "MStractor")
    files <- dir(path, pattern = ".mzXML", full.names = TRUE)
    pd <- data.frame(sample_name = sub(basename(files), pattern = filetype,
        replacement = "", fixed = TRUE), sample_group = SampleGroup,
        stringsAsFactors = FALSE)
    raw_data <- readMSData(files =files, pdata = new("NAnnotatedDataFrame",
        pd), mode = "onDisk")
    ppgExData(raw_data) # example run using ppgExData (see vignettes)
    }
}
