% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakPickingParam.R
\name{PeakPickingParam}
\alias{PeakPickingParam}
\title{Define Peak Picking Parameters}
\usage{
PeakPickingParam()
}
\value{
returns a list input prarmeters
}
\description{
Define and stores peak picking param forthe centWave
    algorithmvia GUI.
}
\details{
The function uses GUIs to define xcms peak picking parameters
    usingthe centwave algorithm. These include:
    minimum and maximum peakwidth,
    minimum and maximum m/z error (in ppm),
    values for the integration threshold
    (set at 5000) and signal to noise threshold,
    the integration method to be used (1 or 2 as defined by the xcms manual)
    and whether the pick picking should fit the gaussian curve.
    The default value for minimum and maximum m/z error and
    signal to noise (S/N) threshold is none and correspond, respectively,
    to 3.03 and 50 ppm and 1000 for S/N value. Detailed information about
    the mentioned parameters can be
    found in the xcms documentation:
    https://www.bioconductor.org/packages/release/bioc/manuals/xcms/
}
\examples{
{MassSpecParam()
    PeakPickingParam()
}
}
\seealso{
\code{\link[xcms]{findChromPeaks-centWave}}
}
