% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PksAnDF.R
\docType{data}
\name{PksAnDF}
\alias{PksAnDF}
\title{An example data frame showing the results of autoCamera function.}
\format{
An object of class data.frame
}
\usage{
data(PksAnDF)
}
\description{
An example data frame showing the results of autoCamera function.
}
\keyword{datasets}
