% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppgExData.R
\name{ppgExData}
\alias{ppgExData}
\title{Perform peak detection and grouping on the example dataset}
\usage{
ppgExData(x)
}
\arguments{
\item{x}{is an OnDisk XCMSnExp raw data-set}
}
\value{
Na
}
\description{
to be used on the example dataset  available within pacakge
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    path<-system.file("extdata",package = "MStractor")
    files <- dir(path, pattern = ".mzXML", full.names = TRUE)
    pd <- data.frame(sample_name = sub(basename(files), pattern = filetype,
        replacement = "", fixed = TRUE), sample_group = SampleGroup,
        stringsAsFactors = FALSE)
    raw_data <- readMSData(files =files, pdata = new("NAnnotatedDataFrame",
        pd), mode = "onDisk")
    ppgExData(raw_data)
}
}
