% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEICs.R
\name{printEICs}
\alias{printEICs}
\title{Prints EICs for XCMSnEXP objects.}
\usage{
printEICs(x, y)
}
\arguments{
\item{x}{is a XCMSnEXP object ('xdata' in thw workflow).}

\item{y}{is a character vector containing the type of EICs to be printed,
possible values are 'raw' and 'filled'.}
}
\value{
Na
}
\description{
prints pngs of the featuresfeature extracted ion chromatograms.
}
\details{
The function is designed for XCMSnEXP objects and has to be used
    forbranch a' of the workflow.Pngs are stored in a dedicated subdirectory
    within the QC folder.
}
\examples{
{ChromParam()
    MassSpecParam()
    PeakPickingParam()
    ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    data(xdata)
    dir.create("./QC")
    printEICs(xdata,'raw')
}
}
\seealso{
\code{\link[xcms]{featureChromatograms}}
}
