% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEICsXset.R
\name{printEICsXset}
\alias{printEICsXset}
\title{Prints EICs for xcmsSet objects}
\usage{
printEICsXset(x, y)
}
\arguments{
\item{x}{is a xcmsSet object.}

\item{y}{is a character vector containing the type of EICs to be printed
The possible values are 'raw' and 'corrected'.}
}
\value{
Na
}
\description{
prints pngs of the feature extracted ion chromatograms.
}
\details{
The function is designed for xcmsSet objects and has to be used
    for branch b' of the workflow.Pngs are stored in a dedicated subdirectory
    withinthe QC folder.
}
\examples{
{ ChromParam()
    MassSpecParam()
    PeakPickingParam()
    ClassType<-c('Mix', 'Treatment1')
    DefineClassAttributes(ClassType)
    data(xSetRef)
    dir.create("./QC")
    printEICsXset(xSetRef,'raw')
}
}
\seealso{
\code{\link[xcms]{featureChromatograms}}
}
