% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{Project}
\alias{Project}
\title{Define MStractor Project}
\usage{
Project()
}
\value{
returns a list containing the working directory and reference file
    paths.
}
\description{
Define the working directory, reference file and QC replicates
    for preliminary evaluation of peak picking
}
\details{
The function does not require arguments. Its execution opens a GUI
    allowing the selection of the working directory
    and 2 reference files chosen
    from one of the folders within the MSfiles directory (e.g.: two Mix
    replicates in the example dataset).
    The function also creates a QC folder
    where graphical outputs to evaluate the progress
    of the workflow are stored.
}
\examples{
{Project()
}
}
\seealso{
\code{\link[base]{getwd}}
}
