% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refTic.R
\name{refTic}
\alias{refTic}
\title{TIC of the reference files.}
\usage{
refTic(x)
}
\arguments{
\item{x}{is the xrefs XCMSnExp object.}
}
\value{
Na
}
\description{
Prints a .png file showing the TIC of the reference files.
}
\details{
The graphical output is stored in the QC folder.
}
\examples{
{  ChromParam()
   MassSpecParam()
   PeakPickingParam()
   SampleGroup<-c("Mix", "Mix")
   symbol<-c(1,1)
   ClassCol<-c("#FF0000FF", "#FF0000FF")
   path<-system.file("extdata",package = "MStractor")
   files <- dir(path, pattern = ".mzXML", full.names = TRUE)
   test<-files[1:2]
   pd <- data.frame(sample_name = sub(basename(test), pattern = filetype,
        replacement ="", fixed = TRUE),sample_group = SampleGroup,
        stringsAsFactors = FALSE)
   ref_data <- readMSData(test, pdata = new("NAnnotatedDataFrame",
        pd), mode = "onDisk")
   expRefData(ref_data)
   dir.create("./QC")
   }
}
\seealso{
\code{\link[xcms]{chromatogram-method}}
}
