% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTalign.R
\name{RTalign}
\alias{RTalign}
\title{Retention Time Alignment on XCMSnEXP objects}
\usage{
RTalign(x, y)
}
\arguments{
\item{x}{is a XCMSnEXP object}

\item{y}{is a character vector defining the correction method
(default is loess)}
}
\value{
xdata, the aligned XCMSnEXP object
}
\description{
Function based on the xcms  'adjustRtime' function.
}
\details{
Retention time correction parameters are entered via GUI.
    The retention time alignment method currently supported is loess
    (see xcms manual).
    A graphical output showing the retetion time deviation is saved
    in the QC folder.
}
\examples{
{ClassType<-c('Mix', 'Treatment1')
DefineClassAttributes(ClassType)
data(xdata)
RTalign(xdata,'loess')
}
}
\seealso{
\code{\link[xcms]{retcor.peakgroups-methods}}
}
