% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTalign_xset.R
\name{RTalign_xset}
\alias{RTalign_xset}
\title{Retention Time Alignment on xcmsSet objects}
\usage{
RTalign_xset(x, y)
}
\arguments{
\item{x}{is a xcmsSet object y is a character vector defining the
correction method (default is loess)}

\item{y}{is a character vector defining the correction method
(default is loess)}
}
\value{
xsAlign, the aligned xcmsSet onbject
}
\description{
Function based on the xcms  'retcor'.
}
\details{
Retention time correction parameters are entered via GUI.
    The retention time alignment method currently supported is loess.
    (see xcms manual).
    A graphical output showing the retetion time deviation is saved in
    the QC folder.
}
\examples{
{data(xSetRef)
    dir.create('./QC')
    RTalign_xset(xSetRef,'loess')
}
}
\seealso{
\code{\link[xcms]{retcor.peakgroups-methods}}
}
