% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsByClass.R
\name{StatsByClass}
\alias{StatsByClass}
\title{Descriptive Stats Calcualtion}
\usage{
StatsByClass(x, y)
}
\arguments{
\item{x}{is the ClassType vector generatedby running the LoadData()
function}

\item{y}{is the xcmsSet}
}
\value{
outputs stored in separate tsv files for each class.
}
\description{
Calculate descriptive stats (Average, St.Dev and CV%) for each
    sample class.
}
\examples{
{data(xSetRef)
    data(BasePksCur)
    MedianNormalize( BasePksCur,xSetRef)
    ClassType<-c('Mix', 'Treatment1')
    ne <- 1
    envir = as.environment(ne)
    assign("ClassType", ClassType, envir)
    StatsByClass(ClassType, xSetRef)
    }
}
