% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testX.R
\docType{data}
\name{testDataSet}
\alias{testDataSet}
\title{HRAM LCMS raw data}
\format{An object of class OnDiskMSnExp}
\usage{
testDataSet
}
\description{
A subset of the example dataset containing 2 high res LCMS raw data files
from the MIX class.
This dataset is used to test the input parameters
}
\keyword{datasets}
