% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata.R
\docType{data}
\name{xdata}
\alias{xdata}
\title{HRAM LCMS  data}
\format{
An object of class XCMSnExp
}
\usage{
data(xdata)
}
\description{
A XCMSnExp object dataset  containing high res LCMS raw data from a wine
metabolomics experiment.
Data are arranged into 2 classes,Treatment 1 and Mix.
Each class has 3 biological replicates. Peak picking and grouping steps
have been already performed.
This dataset is used for testing purposes.
}
\keyword{datasets}
