% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsetConvert.R
\name{xsetConvert}
\alias{xsetConvert}
\title{Converts new xcmsnExp into xcmsset}
\usage{
xsetConvert(x)
}
\arguments{
\item{x}{is a XCMSnExp object to be converted}
}
\value{
Na
}
\description{
XCMSnExp data are converted to xcmsset using the function
    available in the xcms package.
}
\seealso{
\code{\link[xcms]{xcmsSet-class}}
}
