% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawDataSet.R
\docType{data}
\name{rawDataSet}
\alias{rawDataSet}
\title{HRAM LCMS metabolomics experiment}
\format{
An object of class OnDiskMSnExp
}
\usage{
data(rawDataSet)
}
\description{
A dataset containing high res LCMS raw data from a wine metabolomics
experiment.
Data are arranged into 2 classes,Treatment 1 and Mix.
Each class has 3 biological replicates
}
\keyword{datasets}
