% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refTic.R
\name{refTic}
\alias{refTic}
\title{TIC of the reference files.}
\usage{
refTic(x)
}
\arguments{
\item{x}{is the xrefs XCMSnExp object.}
}
\value{
Na
}
\description{
Prints a .png file showing the TIC of the reference files.
}
\details{
The graphical output is stored in the QC folder.
}
\examples{
{  ChromParam()
   MassSpecParam()
   PeakPickingParam()
   data(testX)
   expRefData(testDataSet)
   dir.create('./QC')
   refTic(x_refs)
   }
}
\seealso{
\code{\link[xcms]{chromatogram-method}}
}
